requirements.txt
pymongo
python-nmap